generalized data objects into data_utils.py
